refactor: refactor conflict
