+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     <application
         android:name="com.washingtondcsquad.TudeeApp"
         android:allowBackup="true"
+        android:screenOrientation="unspecified"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
