+++ app/src/main/java/com/washingtondcsquad/tudee/di/serviceModule.kt
@@ -4,11 +4,11 @@ import com.washingtondcsquad.tudee.data.services.AppPreferencesServiceImpl
 import com.washingtondcsquad.tudee.data.services.CategoriesServiceImpl
 import com.washingtondcsquad.tudee.data.services.ImageSaverServiceImpl
 import com.washingtondcsquad.tudee.data.services.TasksServiceImpl
+import com.washingtondcsquad.tudee.domain.provider.StringProvider
 import com.washingtondcsquad.tudee.domain.services.AppPreferencesService
 import com.washingtondcsquad.tudee.domain.services.CategoriesService
 import com.washingtondcsquad.tudee.domain.services.ImageSaverService
 import com.washingtondcsquad.tudee.domain.services.TasksService
-import com.washingtondcsquad.tudee.presentation.provider.StringProvider
 import com.washingtondcsquad.tudee.presentation.provider.StringProviderImp
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
