--- app/src/main/java/com/washingtondcsquad/tudee/presentation/provider/StringProvider.kt
@@ -1,7 +1,0 @@
-package com.washingtondcsquad.tudee.presentation.provider
-
-import androidx.annotation.StringRes
-
-interface StringProvider {
-    fun getString(@StringRes resId: Int): String
-}
