+++ app/src/main/java/com/washingtondcsquad/tudee/presentation/provider/StringProviderImp.kt
@@ -1,6 +1,7 @@
 package com.washingtondcsquad.tudee.presentation.provider
 
 import android.content.Context
+import com.washingtondcsquad.tudee.domain.provider.StringProvider
 
 class StringProviderImp(
     private val context: Context
